+++ src/okio/BufferedSource.java
@@ -21,7 +21,7 @@ public interface BufferedSource extends Source {
 
     short readShortLe() throws IOException;
 
-    void skip(int count) throws IOException;
+    void skip(long count) throws IOException;
 
     int readInt() throws IOException;
